Fixed offline bug which would index invisible characters

+++ indexes.c
@@ -5,7 +5,7 @@
     Date Updated: April 3, 2016
     Description: Contains three index types of indexes, dictionary, postings, and docs.
                  To be used with invertedFileOnline.c
-                 
+
     Tested: No memory leaks or errors
 ***/
 
@@ -14,56 +14,56 @@
 #include "indexes.h"
 #endif
 
-DictIndex initDictIndex(char *term, int df, int postIndex){
+DictIndex initDictIndex(char *term, long df, long postIndex){
     DictIndex index;
-    index.term = malloc(sizeof(char)*(int)strlen(term)+1);
+    index.term = malloc(sizeof(char)*(long)strlen(term)+1);
     index.term = strcpy(index.term, term);
     index.df = df;
     index.postIndex = postIndex;
     return index;
 }
 
-PostIndex initPostIndex(int docno, int tf) {
+PostIndex initPostIndex(long docno, long tf) {
     PostIndex index;
     index.docno = docno;
     index.tf = tf;
     return index;
 }
 
-DocIndex initDocIndex(char *docid, int line) {
+DocIndex initDocIndex(char *docid, long line) {
     DocIndex index;
-    index.docid = malloc(sizeof(char)*(int)strlen(docid)+1);
+    index.docid = malloc(sizeof(char)*(long)strlen(docid)+1);
     index.docid = strcpy(index.docid, docid);
     index.line = line;
     return index;
 }
 
-void printDictArray (DictIndex dIndex[], int dictSize) {
-    for (int i = 0; i < dictSize; i++) {
-        printf("Dictionary: %d: %s %d %d\n", i , dIndex[i].term, dIndex[i].df, dIndex[i].postIndex);
+void printDictArray (DictIndex dIndex[], long dictSize) {
+    for (long i = 0; i < dictSize; i++) {
+        printf("Dictionary: %ld: %s %ld %ld\n", i , dIndex[i].term, dIndex[i].df, dIndex[i].postIndex);
     }
 }
 
-void freeDictArray ( DictIndex dIndex[], int dictSize ) {
-    for (int i = 0; i < dictSize; i++) {
+void freeDictArray ( DictIndex dIndex[], long dictSize ) {
+    for (long i = 0; i < dictSize; i++) {
         free(dIndex[i].term);
     }
 }
 
-void printPostArray (PostIndex pIndex[], int pSize) {
-    for (int i = 0; i < pSize; i++) {
-        printf("Post: %d: %d %d\n", i , pIndex[i].docno, pIndex[i].tf);
+void printPostArray (PostIndex pIndex[], long pSize) {
+    for (long i = 0; i < pSize; i++) {
+        printf("Post: %ld: %ld %ld\n", i , pIndex[i].docno, pIndex[i].tf);
     }
 }
 
-void printDocArray (DocIndex dIndex[], int docSize) {
-    for (int i = 0; i < docSize; i++) {
-        printf("DocNum: %d: %s %d\n", i , dIndex[i].docid, dIndex[i].line);
+void printDocArray (DocIndex dIndex[], long docSize) {
+    for (long i = 0; i < docSize; i++) {
+        printf("DocNum: %ld: %s %ld\n", i , dIndex[i].docid, dIndex[i].line);
     }
 }
 
-void freeDocArray ( DocIndex dIndex[], int docSize ) {
-    for (int i = 0; i < docSize; i++) {
+void freeDocArray ( DocIndex dIndex[], long docSize ) {
+    for (long i = 0; i < docSize; i++) {
         free(dIndex[i].docid);
     }
 }
